requirements/local.txt
# Local development dependencies go here
-r base.txt

coverage==4.4.2
django-coverage-plugin==1.5.0

Sphinx==1.6.5
django-extensions==1.9.7
Werkzeug==0.12.2
django-test-plus==1.0.20
factory-boy==2.9.2

django-debug-toolbar==1.9.1

# improved REPL
ipdb==0.10.3

pytest-django==3.1.2
pytest-sugar==0.9.0


dj-pagination
